requirements.txt
flask
boto3
numpy
arm-pyart
matplotlib
cartopy